add preprocessor script
Shiny has to launch online whenever it's gone idle, so it's important
that we get to launch fast. However, computing Blank/Null votes on
startup was taking quite a bit of time.

We solve that by doing that on a preprocessing script, and just loading
the Shiny app with that preprocessed data.