add "gambiarra" for the 2018 presidential election
As we all know, Lula registered as a candidate in 2018 but didn't get to
run. However, there's a lot of polling data for Lula that can easily get
confusing in a graph when compared with Fernando Haddad's polling data.

So basically what we do is a "gambiarra" to render the result
semi-appropriately: we show Lula polls until the candidate registry
date, and Haddad polls after it. This isn't really that accurate given
how events played out, but it should make for a passable graph.

In any case, we need to think of a better solution on what to do here;
and even if we do choose to stick with this arrangement, the code
definitely needs to improve.